+++ react/components/MinicartIconButton.tsx
@@ -18,8 +18,16 @@ const countCartItems = (
   countMode: MinicartTotalItemsType,
   arr: OrderFormItem[]
 ) => {
-  if (countMode === 'distinctAvailable') {
-    return arr.reduce((itemQuantity: number, item: OrderFormItem) => {
+  if (countMode.indexOf('WithoutAssemblies') !== -1) {
+    // Filter only main products - there is some items with '0' on parentItemIndex
+    arr = arr.filter(item => item.parentItemIndex === null)
+  }
+
+  if (
+    countMode === 'distinctAvailable' ||
+    countMode === 'distinctWithoutAssembliesAvailable'
+  ) {
+    return arr.reduce((itemQuantity: number, item) => {
       if (item.availability === 'available') {
         return itemQuantity + 1
       }
@@ -27,8 +35,11 @@ const countCartItems = (
     }, 0)
   }
 
-  if (countMode === 'totalAvailable') {
-    return arr.reduce((itemQuantity: number, item: OrderFormItem) => {
+  if (
+    countMode === 'totalAvailable' ||
+    countMode === 'totalWithoutAssembliesAvailable'
+  ) {
+    return arr.reduce((itemQuantity: number, item) => {
       if (item.availability === 'available') {
         return itemQuantity + item.quantity
       }
@@ -36,13 +47,13 @@ const countCartItems = (
     }, 0)
   }
 
-  if (countMode === 'total') {
-    return arr.reduce((itemQuantity: number, item: OrderFormItem) => {
+  if (countMode === 'total' || countMode === 'totalWithoutAssemblies') {
+    return arr.reduce((itemQuantity: number, item) => {
       return itemQuantity + item.quantity
     }, 0)
   }
 
-  // countMode === 'distinct'
+  // countMode === 'distinct' || countMode === 'distinctWithoutAssemblies'
   return arr.length
 }