+++ react/typings/global.ts
@@ -5,6 +5,10 @@ type MinicartTotalItemsType =
   | 'distinct'
   | 'totalAvailable'
   | 'distinctAvailable'
+  | 'totalWithoutAssemblies'
+  | 'distinctWithoutAssemblies'
+  | 'totalWithoutAssembliesAvailable'
+  | 'distinctWithoutAssembliesAvailable'
 
 type SlideDirectionType =
   | 'horizontal'